requirements.txt
Flask
Flask-Cors
numpy
Pillow
scikit-learn

